LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/mq-invalid-media-type-layer-002-expected.txt


FAIL 'layer' used as media types is a syntax error assert_equals: expected "not all" but got "not layer"
